Fixed filtering and sorting logic and changed control view
Fixed the sorting and filtering logic. It wasn't allowing filtering without a populated searchbar.

Made it so it can be filtered without anything typed into the searchbar.

Also, updated the controls to only show the searchbar and filter with dropdown.

Also changed the order of them to search then filter