Fixed Sync Problem for checkbox states
The .checkbox-states folder when it synced, because it is a hidden folder as a dot-folder, it would not sync as obsidian doesn't like hidden files.

I tried doing a non dot folder checkbox-states inside of .obsidian folder as that folder does sync, but it only syncs its expected contents and would not sync that folder.

Unfortunately if it isn't a .folder it shows up in your sidebar of folders and files. For now it is named _checkbox-states to put it right at the top.

It auto closes for some reason if you try to open it but this is the least intrusive way I have found to do it.