Made it work with generic column titles
Now when you any generic column title

 test:
    alias: test
    type: string
    searchable: true

filters:
  Familiar:
    test: "$row.Tags?.includes('#test')"

it will save the column title to the JSON file and when you filter, it will work with more than just "Known" and "Prepared"

You can even live change the column title and column fields in the yaml, refresh the doc, then use the checkbox to export data and then try the filter. It works.

Even works with Sync.

If changes are made a full obsidian reload is necessary, not just a document refresh.