Made the logic for filtering checked checkboxes generic
Instead of inserting a #tag into a column that must be named Tags. I made it so it injects the word checked into the same column as the checked checkbox.

This way your filter logic is

    Prepared: "$row.Prepared?.includes('checked')"

so the columnTitle is the column to search and you always have the logic be .includes checked, meaning it will look for the rows in said column that have checked checkboxes (as they will also have the word checked injected (appended) into the same cell in the active memory.

So it doesn't change the document/html, it doesn't break the checkboxes nor change content in reading view breaking the searchbar and filter dropdown