checks checkbox states of all checkboxes upon render
What it was doing before was checking the states of checkboxes that had been interacted with. This meant that any unchecked boxes that were never checked weren't saved.

So if you filtered unchecked boxes, it would only show unchecked boxes that were at one point checked.

Now upon render it properly saves all the checkboxes in the document to the external JSON file