Updated settings tab
Settings tab no longer has that useless button. It now has a button that links to the repo.

I did also add a button for an example YAML and example table that are both in copyable code blocks that you can paste into your doc with appropriate spacing so you can get started.