Updated settings tab and README.md
YAML copy block was missing the ``` that starts and ends the codeblock. Now it has the entire YAML block in one copy button click.

I updated the readme to also have an example of the codeblock and table, but I couldn't get the codeblock to properly copy in the readme.