Optimized Search
When working with large data sets or resource intensive themes, the search function was rather unoptimized. I added a debounce to make it not try to search the entire dataset after every individual keystroke. Instead I added a 200ms delay after the last thing typed to perform a search.

This was in effort to strike a balance for small datasets still having a smooth feeling search and optimizing the search and making it far more convenient to use for large datasets